requirements.txt
bs4==0.0.2
beautifulsoup4==4.13.4
chainlit==2.7.1.1
google-cloud-aiplatform[adk,agent_engines]==1.122.0